Add VR O2 tests to FYI bot
Adds another invocation of chrome_public_test_vr_apk to the VR FYI bot,
but with VrCore set to use the O2 render path. This only works on O MR1+
and has the potential to be less stable, so we only want it on the FYI
bot for now.

Change-Id: I4b528a240eb62d979b586511b223fcac79e66d7c
Reviewed-on: https://chromium-review.googlesource.com/982528
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#547010}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4b897f3566aac85818c960c75741630cb5bfee5d