Don't build all on WebRTC chromium bots.
Our win bot is broken by extensions/common/extension_unittest.cc,
but I can't even find that file in the chromium source tree! Let's try
to get off of 'all' once and for all. Chromium isn't building all for
the most part, so we are sensitive to breakages in obscure targets.

It's not clear if building the remote webapp is enough to prevent
the breakages in http://crbug.com/820041; we failed to repro that one
on our workstations because of nacl build magic, but we didn't
conclude that the bot couldn't have caught the issue since it was
fixed while we were working on it. Well, IIRC.

Let's try this, and if we get more breakages in pnacl_newlib
because of webrtc_overrides problems, let's revisit the problem
and see if we can build some kind of integration test with pnacl
so we detect these errors earlier without building magical targets
(or all for that matter).

Bug: None
Change-Id: Ibc0441ffa17c046f18b5a9b8a82c111f1b03d0f9
Reviewed-on: https://chromium-review.googlesource.com/1011609
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Patrik Höglund <phoglund@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#550570}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a714f9be3c3479e6504329d59c9aa207b12a6b7b