Introduce a protobuf-based FormData fuzzer
As an alternative to the fuzzer from https://crrev.com/c/992496, which
takes an arbitrary string and generates a FormData structure to be
parsed into a PasswordForm, this CL introduces a fuzzer which creates
the FormData structure based on a protobuf decription.

The fuzzer's code is therefore easier to read. The FormData creation
cannot be controlled so tightly as in the case of the string-based
fuzzer, where, e.g., an on-purpose bias is introduced towards empty
field values.

Further, this CL slims down the 'fuzzer_support' target to only
include the unit-tested data_accessor.*. The generator helpers for
both the current and the new, proto-based, fuzzer are bundled inside
the fuzzer_test targets. This allows both only compiling the necessary
files for each fuzzer, and also the need to include the proto_library
as an indirect dependency of components_unittests, which conflicts
with the absence of optimisation for LITE.

The CL also fixes a snippet in the libprotobuf-mutator docs to follow
the style guide, which bans passing non-const references.

Bug: 828705
Change-Id: I89d7753783cc09703a5112bf03bbfdceee81394e
Reviewed-on: https://chromium-review.googlesource.com/1015800
Commit-Queue: Vaclav Brozek <vabr@chromium.org>
Reviewed-by: Jonathan Metzman <metzman@chromium.org>
Reviewed-by: Max Moroz <mmoroz@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#552418}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d42e1353b129a2055636985640682c31b89a9278