cc_perftests: Stop running on the perf bots.
The tests have not been very useful, and has remain disabled on a
number of bots for a while. So stop running the tests on the perf
bots altogether. The tests can still be used locally to watch for
improvements/regressions for specific changes as desired/needed.

BUG=784454, 721757, 736150
BUG=698831, 792995, 723821

Change-Id: Ifede89a3f7a87d96bbb726e0911e1df26e6f9a6b
Reviewed-on: https://chromium-review.googlesource.com/1021653
Reviewed-by: enne <enne@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Ned Nguyen <nednguyen@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#552422}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 87711633bd43436dea338b0869f5ab89c89d758d