Webkit layout test runner: Add option for zero tests executed
Adds --zero-tests-executed-ok, which makes the layout test runner exit
with an ok status if zero tests are executed. This is important for
re-running layout tests without patch, which we do on tryservers to
check if the test was being flaky.

Bug: 533481
Change-Id: I7d6bcd805b99818fdf82ecb88e71863ff72b72e7
Reviewed-on: https://chromium-review.googlesource.com/1018590
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Jeff Carpenter <jeffcarp@chromium.org>
Reviewed-by: Robert Ma <robertma@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#552426}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e3ce42822c4ebe5669dcc33ad9c3b8c77ed1dab1