Android test exclusions: don't exclude boringssl_crypto_tests
These tests do pass on some bots, e.g. [1], and it's not clear why they were
being excluded. If there's a legitimate reason for them not to run on Android,
let's figure it out and get a bug filed.

Sheriffs: if this causes the tests to start failing on Android bots, please
file a bug with information about the error, put the exclusion back and
reference the bug.

Bug: 842698

 [1] https://ci.chromium.org/buildbot/chromium.clang/ToTAndroid/3343

Change-Id: I6e6e5acba61aa3f48bdf4770d6deab998f4723f3
Reviewed-on: https://chromium-review.googlesource.com/1061233
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#559047}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f4e9e18bbaa49760ad46733c8053fa71f88b3742