Android test exclusions: don't exclude google_apis_unittests
These tests do pass on some bots, e.g. [1], and it's not clear why they were
being excluded. If there's a legitimate reason for them not to run on Android,
let's figure it out and get a bug filed.

Sheriffs: if this causes the tests to start failing on Android bots, please
file a bug with information about the error, put the exclusion back and
reference the bug.

(This also enables it on a Windows and some Cast bots; it probably passes
on those too.)

Bug: 842698

 [1] https://ci.chromium.org/buildbot/chromium.clang/ToTAndroid/3343

R=kbr@chromium.org, thakis@chromium.org

Change-Id: I28c59b515fbbd757783ab7c8f54f84557b9c0c67
Reviewed-on: https://chromium-review.googlesource.com/1063729
Commit-Queue: Nico Weber <thakis@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#559514}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9b822f03e769c09a8c22a20cd26bcd0386aeb245