Android test exclusions: don't exclude viz_unittests
These tests do pass on some bots, e.g. [1], and it's not clear why they
were being excluded. If there's a legitimate reason for them not to run
on Android, let's figure it out and get a bug filed.

Sheriffs: if this causes the tests to start failing on Android bots,
please file a bug with information about the error, put the exclusion
back and reference the bug.

Bug: 842698

 [1] https://ci.chromium.org/buildbot/chromium.clang/ToTAndroid/3343

R=thakis@chromium.org

Change-Id: I9a5cdf61e6b001cf8dae6cab3eee457cf74bbf76
Reviewed-on: https://chromium-review.googlesource.com/1068741
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#560980}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 34037a3b48a18a65286e80247ae7f9b82d02ea01