Run chrome_app_unittests on the non-android memory and clang tot bots.
It was already running on the win clang tot bots, but not yet on
the mac and linux bots.

This is part of an effort to reduce the number of test configurations.
Originally reviewed as part of https://chromium-review.googlesource.com/1061795,
which I'm now landing in smaller pieces.

TBR=kbr

Bug: 843511
Change-Id: Ieffbc32b066c25d9626652cbd47c7074efd0e81c
Reviewed-on: https://chromium-review.googlesource.com/1080257
Commit-Queue: Nico Weber <thakis@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#563275}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0b1ae7ea0388fbbf2cc76def9917e3ab5f54a444