Run nacl_loader_unittests on the clang tot bots.
This is part of an effort to reduce the number of test configurations.
Originally reviewed as part of https://chromium-review.googlesource.com/1061795,
which I'm now landing in smaller pieces.

In the reland of this part, I'm adding a new exception for the mac asan bot.
The mac asan bot has nacl disabled, and the original change turned the bot red:
https://ci.chromium.org/buildbot/chromium.memory/Mac%20ASan%2064%20Builder/62745

TBR=kbr

Bug: 843511
Change-Id: Idf302a474570ab8d3e25858321470898e8387f4d
Reviewed-on: https://chromium-review.googlesource.com/1079846
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#563278}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 05d3e64b75533286442c1daa0d331132959f37ed