Run sync_integration_tests on the mac asan memory and the linux and win clang tot bots.
(It's suppressed on the non-mac memory bots for unknown reasons.
It's already running and on the mac clang tot bots.)

This is part of an effort to reduce the number of test configurations.
Originally reviewed as part of https://chromium-review.googlesource.com/1061795,
which I'm now landing in smaller pieces.

Bug: 843511
Change-Id: Id2968707671d7edda5f9c9cacee5c747d86dfff7
Reviewed-on: https://chromium-review.googlesource.com/1080260
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#563307}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f12034993a38dc665ca0d92fada79fcc87ac8a71