chromeos: adds support for RequestClose and wires up in ash
When a remote client creates a window, the remote client should control when it
is closed. This makes it to so that when the user attemps to close a window
created by a remote client, the remote client is asked to close the window.

BUG=842298
TEST=covered by test

Change-Id: I9ace655c0648fc0f8943425c67fb57e088ef00fa
Reviewed-on: https://chromium-review.googlesource.com/1087870
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Michael Wasserman <msw@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#564897}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 94392859d28397a3b0260fd7add1e776d965e9dc