mac: Let memory and tot bots use the same test list as the main waterfall bots.
This ensures that the memory and clang tot bots run the same tests as the
main waterfall bots.

Everything running on the main waterfall should also be tested under ASan,
and it must be tested with clang/tot too to make sure compiler updates don't
break anything. This change makes it impossible for things to get out of sync
(on mac).

If https://chromium-review.googlesource.com/c/chromium/src/+/1088754 hadn't
been reverted, this would've been behavior-preserving. As-is, it adds
headless_browsertests and headless_unittests to the mac asan bot and the
clang tot bots (which is fine).

Bug: 843511
Change-Id: I5e3cea35d99d630d7eb41154984abc23defe6219
Reviewed-on: https://chromium-review.googlesource.com/1089136
Commit-Queue: Nico Weber <thakis@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#564974}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fae3606bf9174a0f0ae0304dc91d2cd6d74fa819