Force gpu=none on quarantined mac browser_tests, remove from 10.10.
This sets the gpu=none flag explicitly to get the browser_tests to
run on the VMs in the Chrome-quarantine pool on Mac10.12, to avoid any
confusion with the default gpu values for the other test suites.

This also temporarily turns off the browser_tests on 10.10, since it
looks like we're probably too low on capacity there to handle this
easily. We'll add more tomorrow and try again.

TBR=jbudorick@chromium.org
BUG=828031
NOTRY=true

Change-Id: Ib948ba17c2bb3b4fb56912a9c612e72fa5e08f11
Reviewed-on: https://chromium-review.googlesource.com/1103938
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#567933}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9e8bb711222aaf2af8c77934b0c78532d85cf5d6