Update SitePerProcessHitTestBrowserTest.HitTestStaleDataDeletedView hit test waiting
Currently SitePerProcessHitTestBrowserTest.HitTestStaleDataDeletedView is using
a legacy method for waiting for hit test data. This change updates it to use
the new HitTestRegionObserver.

However there is a bug in the /2 variant of the test, there overlapping hit test
regions are not submitted. Due to this I have it exit early in that config.

TEST=SitePerProcessHitTestBrowserTest.HitTestStaleDataDeletedView

Bug: 846798
Change-Id: I410d493f1abc7334e4cc21c3c4edceb4e2edd8c9
Reviewed-on: https://chromium-review.googlesource.com/1102721
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#568447}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 79f35d0e90605e72245b8c1ea5d560e5f9e0b6ab