Monochrome: Ignore isolated script test argument when retrying
This argument is passed when the test is being retried. Parse and ignore
it so the retry without patch logic works correctly.

Also fixes the json test result output for the test, so that if the test
fails, it's marked as unexpected.

Also moves the wrapper script into //testing/scripts, which is where
wrapper scripts are supposed to live.

Bug: 533481
Change-Id: I4409be32a84c885d677ae90791a51c847d0c7e1f
Reviewed-on: https://chromium-review.googlesource.com/1107054
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Mohamed Heikal <mheikal@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#569510}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e58ad2f781177df53a4a0ea8a81847bf9e71429a