Remove test exceptions for 'Win 10 Fast Ring' bot.
The currently disabled set of tests on that bot is fairly arbitrary. Since
the bot is currently red anyway and has been for a while (see 2nd bug), let's
run all tests there, and when we get around to greening it up we can then
disable suites that turn out to be unamenable to fixing. It likely won't be
the suites that are currently disabled.

Bug: 843511,851705
Change-Id: Ifabc504fe56ebdbdab32609055cdaf39d8d73d11
Reviewed-on: https://chromium-review.googlesource.com/1120227
Commit-Queue: Nico Weber <thakis@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#571571}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a8e6d828e89e396c1fc9e0e2dc78c206adbd2a06