[VK] Move keyboard to focused display if no display is touchable.
Currently when no display is touchable, we show the virtual keyboard
on the primary display. According to https://crbug.com/297858, this
was just a basic heuristic to get it working. As mentioned in
https://crbug.com/303429#c5, forcing the user to change their primary
display all the time is not very user friendly.

We change this so that when no display is touchable, we show the
keyboard on the focused window (if any).

Bug: 856613
Change-Id: I4e0524747c54e471607b2b546a3ef663ad215b55
Reviewed-on: https://chromium-review.googlesource.com/1118060
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#571822}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 019eb9605d08b14a5a796c8e8071fd96775b3a02