Update TouchInputBrowserTest Frame Synchronizing
The TouchInputBrowserTests were flaking across OSes.
This was due to how it was waiting for frames. It was waiting for a
visual properties synchronization to have been started. However, while
this leads to a frame being generated, it does not block until that
frame is produced.

This change updates the frame syncing.
Now it will use RenderFrameSubmisisonObserver to wait for a frame from the URL load to be generated.
For non-Android it will then set a size, which triggers a visual properties sync. Then it will wait for a frame to be generated.
On Android, this size is set in the initial page load, so there is no synchronizing here.
Furthermore for Viz Hit Testing, a wait for hit test data is added.

TEST=TouchInputBrowserTest

Bug: 346363, 514456,855532
Change-Id: Ifa9179af9234d98c5be5575876c061763edb9f4e
Reviewed-on: https://chromium-review.googlesource.com/1147322
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#578819}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9db0a14b711415d66efdcec8eb88971a9b2efdb0