Fixes ExtensionWebRequestApiTest.WebRequestTypes with network service
Tests now wait for the webRequest listeners to be set up, and the
correct render_frame_id is now used in WebRequestInfo.

The reason tests need to wait for listeners to be set up with the
network service is that requests need to be proxied through
WebRequestProxyingURLLoaderFactory for web request listeners to be
called. Since this proxy is in the browser process, we only set it
up when necessary for performance reasons. This means that if there
are no web request listeners set up, we do not proxy requests, but
when the first listener is set up we then reset all
URLLoaderFactories, which causes them to be recreated with the
proxy. Because of this, there is a behavior difference in the
network service, where code like:

  chrome.webRequest.foo.addListener(...);
  sendSomeRequest(...);

will no longer work with network service. This should probably be
fine since this is a very rare occurrence. For more discussion on
this, see crbug.com/784576#c19.

Bug: 721414
Cq-Include-Trybots: luci.chromium.try:linux_mojo
Change-Id: I92fb4a02b14641886624a726301acc9e95f53533
Reviewed-on: https://chromium-review.googlesource.com/1149407
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Ken Rockot <rockot@chromium.org>
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#579115}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1352fa135bc3a3b28d27ebc81519727831a5ba08