Explicitly target GPUs when falling back to mac10.12 for layout_tests.
I think this is why the fallback isn't working at all:
https://chromium-swarm.appspot.com/tasklist?c=name&c=state&c=created_ts&c=duration&c=pending_time&c=pool&c=bot&et=1534289040000&f=buildername-tag%3Amac_chromium_rel_ng&f=name-tag%3Awebkit_layout_tests&f=os%3AMac-10.12&l=50&n=true&s=created_ts%3Adesc&st=1534202640000

It's inheriting the "gpu:none" that the recipe default to:
https://codesearch.chromium.org/chromium/build/scripts/slave/recipe_modules/swarming/api.py?rcl=48ebe7c7e93d140e1746c034fbf39d2480bd34f0&l=188

And we don't have any 10.12 VMs:
https://chromium-swarm.appspot.com/botlist?c=id&c=os&c=task&c=status&f=os%3AMac-10.12&f=pool%3AChrome&f=gpu%3Anone&l=100&q=gpu%3Anone&s=id%3Aasc

Hopefully this helps things. If it doesn't, I'm going to move the
entire suite back over to 10.12. We simple don't have the capacity for
this... and won't for a while.

Bug: 854874
Change-Id: I87f218a932266a0eef311741f20209f4ece0f8f8
Reviewed-on: https://chromium-review.googlesource.com/1174919
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#583355}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4ef635e8b9d55aecad9ec23db4e8ebd103984ad9